requirements.txt
streamlit>=1.20.0
altair<5
streamlit-keyup
spacy
openai
python-dotenv
aiohttp